Properly implemented graphics
Refactoring and rustfmt preformed.

New thread now preformes window updates.
On update events: the system will check for frames from the cpu,
currently a frame is every pixel given along with its spot on the
screen.  Once handled, the frame is saved to a texture and displayed.
If the update took too long, then fallback will be not to draw a frame,
but continue to handle window events, and try again next update.